(PC-4626) move: organize files in lib
